design: 투표 화면 css
