+++ src/components/Voting.tsx
@@ -3,22 +3,15 @@
 import { Food } from '@/types/party';
 import { useSession } from 'next-auth/react';
 import VotingForm from './VotingForm';
+import FoodCard from './FoodCard';
 
 type Props = {
   foods: Food[];
-  partyId:string
 };
 
-const Voting = ({ foods,partyId }: Props) => {
+const Voting = ({ foods }: Props) => {
   const { data: session } = useSession();
   const user = session?.user;
-
-  const onClickVoteButton = (foodId: string) => {
-    fetch('/api/vote', {
-      method: 'PUT',
-      body: JSON.stringify({ foodId }),
-    }).then((res) => res.json());
-  };
 
   const checkVoted = () => {
     const voters = foods.flatMap((food) => food.voters);
@@ -30,18 +23,10 @@ const Voting = ({ foods,partyId }: Props) => {
       <ul>
         {foods.map((food) => (
           <li key={food.id} className='flex gap-3'>
-            <p>{food.name}</p>
-            {food.voters && <p>{food.voters.length}</p>}
-            <button
-              onClick={() => onClickVoteButton(food.id)}
-              disabled={checkVoted()}
-            >
-              투표
-            </button>
+            <FoodCard disabled={checkVoted()} food={food} />
           </li>
         ))}
       </ul>
-      <VotingForm partyId={partyId}/>
     </div>
   );
 };
